requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.17
requests==2.32.3
anthropic==0.40.0
langchain==0.3.9
langchain-anthropic==0.2.4
langchain-community==0.3.8
langchain-core==0.3.19
twilio==9.3.7
deepgram-sdk==3.8.3
pydub==0.25.1
httpx==0.28.1
python-dotenv==1.0.1
cartesia==2.0.9
aiofiles==23.2.1
